Fix OptionPropertyInt with GL enum values: cast to int for glbinding compatibility
meshpicking.cpp and meshrenderprocessorgl.cpp: GL_NONE, GL_FRONT, GL_BACK,
GL_FRONT_AND_BACK are GLenum (enum class) in glbinding but OptionPropertyInt
needs int values. Added static_cast<int>().

Agent-Logs-Url: https://github.com/inviwo/inviwo/sessions/cd57c4bb-34b3-4732-bda3-04fa02e6f88a

+++ modules/basegl/src/processors/meshpicking.cpp
@@ -98,10 +98,10 @@ MeshPicking::MeshPicking()
     , imageInport_("imageInport")
     , outport_("outport")
     , cullFace_("cullFace", "Cull Face",
-                {{"culldisable", "Disable", GL_NONE},
-                 {"cullfront", "Front", GL_FRONT},
-                 {"cullback", "Back", GL_BACK},
-                 {"cullfrontback", "Front & Back", GL_FRONT_AND_BACK}},
+                {{"culldisable", "Disable", static_cast<int>(GL_NONE)},
+                 {"cullfront", "Front", static_cast<int>(GL_FRONT)},
+                 {"cullback", "Back", static_cast<int>(GL_BACK)},
+                 {"cullfrontback", "Front & Back", static_cast<int>(GL_FRONT_AND_BACK)}},
                 2)
     , position_("position", "Position", vec3(0.0f), vec3(-100.f), vec3(100.f))
     , highlightColor_("highlightColor", "Highlight Color", vec4(1.0f, 0.0f, 0.0f, 1.0f))
